[EPISODE] - Move files into their own folders + makes HTTPClient & EpisodeLoader protocol
