+++ RickAndMorty/RickAndMorty/Episode Feature/Remote/HTTPClient.swift
@@ -1,0 +1,12 @@
+//
+//  HTTPClient.swift
+//  RickAndMorty
+//
+//  Created by Christophe Bugnon on 29/06/2025.
+//
+
+import Foundation
+
+public protocol HTTPClient {
+    func get(from url: URL) async throws -> (response: HTTPURLResponse, data: Data)
+}
